requirements.txt
Pillow
pytesseract
pdf2image
EasyOCR
opencv-python # EasyOCR often uses OpenCV
pytest 
pathlib
PyMuPDF
google-generativeai
python-dotenv